i2c: qup: Use the interconnect API
The interconnect API provides an interface for consumer drivers to express
their bandwidth needs. This data is then aggregated and the interconnect
hardware is configured to the most appropriate power/performance profile.

Use the interconnect API to get() the path between the endpoints used for
data transfers by the I2C QUP and report the needed bandwidth.

Signed-off-by: Georgi Djakov <georgi.djakov@linaro.org>