interconnect: Add support for path tags
Consumers may have use cases with different bandwidth requirements based on
the system or driver state. The consumer driver can append a specific tag to
the path and pass this information to the interconnect platform driver to do
the aggregation based on this state.

Introduce icc_set_tag() function that will allow the consumers to optionally
append tag to each path. The aggregation of these tagged paths is platform
specific.

Signed-off-by: Georgi Djakov <georgi.djakov@linaro.org>