overlays: add better target to rpi-ft5406
The current overlay's target is set to "/firmware". This is worng as the
real firmware node is located at "/soc/firmware". This is not ideal as
the touchcreen didn't appear in the final device tree as a child of the
firmware node.

the overlay still worked fine as the driver uses 'of_get_parent()' to
get the firmware node which in the end pointed to the right place.

Signed-off-by: Nicolas Saenz Julienne <nsaenzjulienne@suse.de>